aceptar solicitud casi andando

+++ easyFlete/src/main/java/com/grupo15/BusinessLogic/IBLSolicitud.java
@@ -20,5 +20,5 @@ public abstract interface IBLSolicitud {
     public String calcularPrecio(String solicitud, String fletero);
     public String getSolicitudGEO(int id);
 
-    public boolean aceptarSolicitud(int solicitud, String fletero, int precio);
+    public boolean aceptarSolicitud(int solicitud, String fletero, double precio);
 }
